refactor: Split tc.rs for each command
