+++ src/cli/tc/init.rs
@@ -1,0 +1,95 @@
+use std::fs;
+use std::path::Path;
+
+use anyhow::Result;
+use inquire::Confirm;
+use stripmargin::StripMargin;
+
+use super::helpers::{is_exists, FILE_NAME, JSON_FILE_NAME, YML_FILE_NAME};
+
+pub(super) async fn init_command(folder_path: &str) -> Result<()> {
+    let folder_exists = is_exists(folder_path);
+
+    if !folder_exists {
+        let confirm = Confirm::new(&format!(
+            "Do you want to create the folder {}?",
+            folder_path
+        ))
+        .with_default(false)
+        .prompt()?;
+
+        if confirm {
+            fs::create_dir_all(folder_path)?;
+        } else {
+            return Ok(());
+        };
+    }
+
+    let tailcallrc = include_str!("../../../generated/.tailcallrc.graphql");
+    let tailcallrc_json: &str = include_str!("../../../generated/.tailcallrc.schema.json");
+
+    let file_path = Path::new(folder_path).join(FILE_NAME);
+    let json_file_path = Path::new(folder_path).join(JSON_FILE_NAME);
+    let yml_file_path = Path::new(folder_path).join(YML_FILE_NAME);
+
+    let tailcall_exists = fs::metadata(&file_path).is_ok();
+
+    if tailcall_exists {
+        // confirm overwrite
+        let confirm = Confirm::new(&format!("Do you want to overwrite the file {}?", FILE_NAME))
+            .with_default(false)
+            .prompt()?;
+
+        if confirm {
+            fs::write(&file_path, tailcallrc.as_bytes())?;
+            fs::write(&json_file_path, tailcallrc_json.as_bytes())?;
+        }
+    } else {
+        fs::write(&file_path, tailcallrc.as_bytes())?;
+        fs::write(&json_file_path, tailcallrc_json.as_bytes())?;
+    }
+
+    let yml_exists = fs::metadata(&yml_file_path).is_ok();
+
+    if !yml_exists {
+        fs::write(&yml_file_path, "")?;
+
+        let graphqlrc = r"|schema:
+         |- './.tailcallrc.graphql'
+    "
+        .strip_margin();
+
+        fs::write(&yml_file_path, graphqlrc)?;
+    }
+
+    let graphqlrc = fs::read_to_string(&yml_file_path)?;
+
+    let file_path = file_path.to_str().unwrap();
+
+    let mut yaml: serde_yaml::Value = serde_yaml::from_str(&graphqlrc)?;
+
+    if let Some(mapping) = yaml.as_mapping_mut() {
+        let schema = mapping
+            .entry("schema".into())
+            .or_insert(serde_yaml::Value::Sequence(Default::default()));
+        if let Some(schema) = schema.as_sequence_mut() {
+            if !schema
+                .iter()
+                .any(|v| v == &serde_yaml::Value::from("./.tailcallrc.graphql"))
+            {
+                let confirm =
+                    Confirm::new(&format!("Do you want to add {} to the schema?", file_path))
+                        .with_default(false)
+                        .prompt()?;
+
+                if confirm {
+                    schema.push(serde_yaml::Value::from("./.tailcallrc.graphql"));
+                    let updated = serde_yaml::to_string(&yaml)?;
+                    fs::write(yml_file_path, updated)?;
+                }
+            }
+        }
+    }
+
+    Ok(())
+}
