+++ src/cli/tc/start.rs
@@ -1,0 +1,18 @@
+use anyhow::Result;
+
+use super::cli::fmt::Fmt;
+use super::cli::server::Server;
+use super::helpers::log_endpoint_set;
+use crate::core::config::reader::ConfigReader;
+
+pub(super) async fn start_command(
+    file_paths: Vec<String>,
+    config_reader: &ConfigReader,
+) -> Result<()> {
+    let config_module = config_reader.read_all(&file_paths).await?;
+    log_endpoint_set(&config_module.extensions.endpoint_set);
+    Fmt::log_n_plus_one(false, &config_module.config);
+    let server = Server::new(config_module);
+    server.fork_start().await?;
+    Ok(())
+}
